requirements.txt
astroid==2.15.6
colorama==0.4.6
dateutils==0.6.12
dill==0.3.7
et-xmlfile==1.1.0
greenlet==2.0.2
isort==5.12.0
lazy-object-proxy==1.9.0
loguru==0.7.0
mccabe==0.7.0
metaclass==1.3
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
pandas-stubs==2.0.3.230814
platformdirs==3.10.0
psycopg2==2.9.7
pydantic==1.10.12
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
sqlalchemy2-stubs==0.0.2a35
tomlkit==0.12.1
types-python-dateutil==2.8.19.14
types-pytz==2023.3.0.1
typing_extensions==4.7.1
tzdata==2023.3
win32-setctime==1.1.0
wrapt==1.15.0
